+++ lab-2/src/jvmMain/kotlin/io/github/e1turin/output/view/entities/settings/model/SimpleIterationSystemSettings.kt
@@ -5,7 +5,7 @@ import com.arkivanov.decompose.value.Value
 import io.github.e1turin.output.view.shared.lib.decompose.mutate
 
 
-class SystemSimpleIterationSettings : SystemSettings {
+class SimpleIterationSystemSettings : SystemSettings {
     private val _data = MutableValue(
         SystemSimpleIterationData(
             range = listOf(
