+++ lab-2/src/jvmMain/kotlin/io/github/e1turin/output/view/entities/task/model/Task.kt
@@ -34,6 +34,6 @@ private inline fun mapEquationSettings(taskType: TaskType.Equation) = when (task
 }
 
 private inline fun mapSystemSettings(taskType: TaskType.System) = when (taskType) {
-    TaskType.System.SimpleIteration -> SystemSimpleIterationSettings()
+    TaskType.System.SimpleIteration -> SimpleIterationSystemSettings()
     TaskType.System.Undefined -> DefaultSystemSettings
 }
