+++ lab-2/src/jvmMain/kotlin/io/github/e1turin/output/view/features/select_settings/ui/SettingsPanel.kt
@@ -29,7 +29,7 @@ fun SettingsPanel(
     ) {
         when (model) {
             is NewtonEquationSettings -> NewtonSettingsForm(modifier = Modifier, settings = model)
-            is SystemSimpleIterationSettings -> SystemSimpleIterationForm(modifier = Modifier, settings = model)
+            is SimpleIterationSystemSettings -> SystemSimpleIterationForm(modifier = Modifier, settings = model)
             //TODO: another methods
 
             is DefaultSettings -> Box(contentAlignment = Alignment.Center) {
